[tests] Run tests in parallel
Context: TODO!

Commit 91a39725 updated `make run-all-tests` so that certain MSBuild
unit test targets would be run in parallel, but had two flaws:

 1. It only worked on non-Windows platforms.
 2. It ignored many errors.

Commit 91a39725 only worked on non-Windows platforms because it worked
by generating a **bash**(1)-compatible shell script which
*backgrounded* the target invocations.  Errors were ignored because
the generated script always exited with an error code of 0:

	msbuild RunTests.targets /t:RunNUnitTests &
	msbuild RunTests.targets /t:RunApkTests &
	# ...
	exit 0

Such a script won't easily run on Windows, and the `msbuild` exit
codes were ignored because I couldn't think of a good and "reasonable"
way to capture them.  (To be fair, Google/StackOverflow/etc. did show
various ways to capture the exit codes for backgrounded pipelines.
I considered them all to be largely *unreadable*.)

Additionally, I didn't think that this would necessarily be a problem:
if a unit test failed, it should report the failure in one of the
NUnit `.xml` files -- largely so that we wouldn't need to read the
Jenkins Console Output to see what the *actual* error was! -- and thus
ignoring the exit status shouldn't be a problem.

Turns Out...that's wrong.  We have observed builds which are "green"
-- no errors! -- but in which there *were* failures, and the failures
weren't reported *because* the `parallel-targets.sh` script didn't
report that anything was wrong!

TODO: Insert description of the crash/failure/whatever which was
actually ignored.  Requires finding a suitable build log. Film at 11.

Doh!

Fix both issues: since sanely capturing background process exit codes
is not "reasonably easy" to do in **bash**(1), *Don't Use Bash*.
Instead, introduce a new `<RunParallelTargets/>` task which uses
`System.Diagnostics.Process` to run the targets in the background,
then (sanely) wait for them to complete and perform any necessary
error checking.

This has the added benefit that it *also* now works on Windows.